+++ src/ui/src/app/index.ts
@@ -1,5 +1,4 @@
 export * from "./bootstrap.js";
 export * as Components from "./components/index.js";
 export * from "./contracts.js";
-export * from "./graph-availability.js";
 export * as State from "./state/index.js";
