+++ src/ui/test/graph-availability.test.ts
@@ -2,7 +2,7 @@ import assert from "node:assert/strict";
 import test from "node:test";
 
 import type { GraphVisualizationUiModel } from "../src/app/contracts.js";
-import { hasGraphEdges, hasLoadedGraphIndex, hasLoadedGraphProject } from "../src/app/graph-availability.js";
+import { hasGraphEdges, hasLoadedGraphIndex, hasLoadedGraphProject } from "../src/app/contracts.js";
 
 function createUiModel(): GraphVisualizationUiModel {
     return {
